rename operators name for kbc (#79)
