+++ dataflow/operators/generate/KnowledgeCleaning/__init__.py
@@ -1,7 +1,7 @@
-from .CorpusTextSplitter import CorpusTextSplitter
-from .KnowledgeExtractor import KnowledgeExtractor
-from .KnowledgeCleaner import KnowledgeCleaner
-from .MultiHopQAGenerator import MultiHopQAGenerator
+from .corpus_text_splitter import CorpusTextSplitter
+from .knowledge_extractor import KnowledgeExtractor
+from .knowledge_cleaner import KnowledgeCleaner
+from .multihop_qa_generator import MultiHopQAGenerator
 
 __all__ = [
     "CorpusTextSplitter",
